[Feature] Add the metrics for the number of total partitions (#24)
### What changes were proposed in this pull request?
Add a metrics that represent the number of total partitions in a server.


### Why are the changes needed?
The total number of  partitions is useful for us. It can reflect the load of the shuffle server. If a shuffle server has too many partitions, the shuffle server will cause too much random io, the shuffle server will have poor performance.


### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Manual test and ut

Co-authored-by: roryqi <roryqi@tencent.com>