[Feature] Read handler refactor (#34)
### What changes were proposed in this pull request?
1. Add shuffle read  handlers including hdfs and multistorage.
2. Tidy the file readers to make them cohesive.
3. Unify the read api used by client read handlers including hdfs and multistorage.

### Why are the changes needed?
We are refactoring the reading process to read the index first and then the data. So we need handlers to handle shuffle read logic that read index files, parse them to data segment and then read data by offset and length, in which the index and data of the shuffle are fetched from servers or read directly from hdfs. 
There may be multiple files for a single partition and we read the file pair (index and data) one by one so one shuffle read handler would only deal a pair, we will list and create multiple shuffle-read handlers in the client handler that  only contain a pair of shuffle index and data file readers.


### Does this PR introduce _any_ user-facing change?
No


### How was this patch tested?
ut, integration test.

Co-authored-by: macduan <macduan@tencent.com>