[Feature] Support StorageType MEMORY_LOCALFILE_HDFS (#63)
### What changes were proposed in this pull request?
We implement the storage MEMORY_LOCALFILE_HDFS.

### Why are the changes needed?
We hope use memory, localfile and hdfs at the same time. and i find hdfs read handler need to be created lazily. If not, we could find nay index in the hdfs, but data could be flushed to hdfs when data is reading.

### Does this PR introduce _any_ user-facing change?
User can use a new storageType MEMORY_LOCALFILE_HDFS for `rss.storage.type`

### How was this patch tested?
Add new uts.

Co-authored-by: roryqi <roryqi@tencent.com>