[Feature] Firestorm supports MapReduce (Part 2 - Map Side) (#108)
### What changes were proposed in this pull request?
We add a new MapOutputBuffer to support the use of rss. Actually MapOutputBuffer is a shuffle writer. MapOutputBuffer send data to Shuffle Server.

### Why are the changes needed?
In real production environments, there are too many MapReduce's jobs. The work is meaningful.

### Does this PR introduce _any_ user-facing change?
Yes, we bring some new configurations. In another pr, we will update the doc about the use of this feature.

### How was this patch tested?
new Uts.

Co-authored-by: roryqi <roryqi@tencent.com>