[Bugfix] Map side avoid producing too many blocks (#161)
### What changes were proposed in this pull request?
We change the sort order, we should send bigger size block first, otherwise we will produce too many blocks. And it will take too much time to send them. And we increase sendThreshold parameter, if we don't have this parameter, when we have much in-send memory, we will send small buffer frequently.

### Why are the changes needed?
If we don't have this patch,  the task will fail because of timeout.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
New ut

Co-authored-by: roryqi <roryqi@tencent.com>