Support using remote fs path to specify the excludeNodesFilePath (#200)
What changes were proposed in this pull request?
Support using remote fs path to specify the excludeNodesFilePath

Why are the changes needed?
When existing two coordinators serving for online, we hope they can read the consistent exclude nodes file insteading of using the localfile syncing manually.

Does this PR introduce any user-facing change?
Yes. It's an incompatible change.

When the default fs is HDFS in the core-site.xml, and the excludeFilePath is specified to "/user/xxxxx" in coordinator server.
After applied this patch, filesystem will be initialized to remote HDFS due to lacking scheme.

How was this patch tested?
Unit tests.