[Chore]: document update and build optimize
When building uniffle locally, I noticed that it requires JDK1.8 to build and suspicious
maven-default-http-blocker for json-smart.
```
[WARNING] net.minidev:json-smart/maven-metadata.xmlfailed to transfer from http://0.0.0.0/ during a previous attempt. This failure was cached in the local repository and resolution will not be reattempted until the update interval of maven-default-http-blocker has elapsed or updates are forced. Original error: Could not transfer metadata net.minidev:json-smart/maven-metadata.xml from/to maven-default-http-blocker (http://0.0.0.0/): transfer failed for http://0.0.0.0/net/minidev/json-smart/maven-metadata.xml
```

To address these two issues, README.md is updated to reflect the requirement. The json-smart dependency
is excluded from hadoop-common since it's never used.

Before I'm making a PR, I found the project name in CONTRIBUTING.md is not changed yet. Thus, this commit
also updates the CONTRIBUTING.md and makes some amend to the doc.