[improvement] Remove experimental feature with ShuffleUploader (#51)
### What changes were proposed in this pull request?
ShuffleUploader is an experimental feature which is target to merge and upload data from local disk to remote storage if local disk hasn't enough space. It is replaced by `MEMORY_LOCAL_HDFS` and should be removed to avoid maintenance in the future.


### Why are the changes needed?
The experimental feature is redundancy and we have to maintain it during update with new features about storage. 

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
with exist UTs