[Improvement] Should match from pathToStorages when appId does not exist in appIdToStorages (#168)
### What changes were proposed in this pull request?
![createlog](https://user-images.githubusercontent.com/84573424/185648279-a6a51d39-cd1d-4b0a-8c2a-95069fb96cd2.png)
![deletelog](https://user-images.githubusercontent.com/84573424/185648346-b320fce4-7131-404b-baaf-4def5ad7a1a1.png)
From the audit log of HDFS, it can be seen that when the HDFS path of this app was last deleted at 18:00:55, the log in the `shuffleServer` found that the error about `file could not be found`, and the file would continue to be written. At last we found that when some appId cache was removed in `appIdToStorages`, and then `HdfsStorageManager` calls `removeResources` will cause storagePath to not be deleted.

### Why are the changes needed?
When the cache of `appIdToStorages` removed, the remote path can be deleted normally.
### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Added ut.