[Improvement][AQE] Avoid calling getShuffleResult multiple times (#190)
###What changes were proposed in this pull request?
For issue #136  , When we use AQE, we may call shuffleWriteClient.getShuffleResult multiple times. But if both partition 1 and partition 2 are on the server A, we call getShuffleResult(partition 1) to get data form server A, and then we call getShuffleResult(partition 2) to get data form server A, it's not necassray. We can get getShuffleResult(partition 1, partition 2) instead.

###Why are the changes needed?
Improve getShuffleResult

###Does this PR introduce any user-facing change?
No

###How was this patch tested?
Added UT

Co-authored-by: leixianming <leixianming@didiglobal.com>