Set the default disk capacity to the total space (#237)
### What changes were proposed in this pull request?
Set the default disk capacity to the total space

### Why are the changes needed?
When shuffle-servers are colocated with NodeManagers, the disk capacity is hard to set. Because when it's equal with the total space, it will cause the most shuffle-servers startup fail due to free-space checking. If not, it will waste some disk resources.

And the default disk capacity is 1TB, it's not out-of-box value. So this PR is to set the default value of 'rss.server.disk.capacity' to the disk total space.


### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
No need.