[ISSUE-273][BUG] Get shuffle result failed caused by concurrent calls to registerShuffle (#274)
What changes were proposed in this pull request?
For issue#173, make the lifecycle of remoteStorage var scoped in the method of registerShuffle

Why are the changes needed?
This problem causes getShuffleResult to fail, which eventually causes the task to fail.

Does this PR introduce any user-facing change?
No

How was this patch tested?
Existing UT

Co-authored-by: leixianming <leixianming@didiglobal.com>