[BUG][AQE][LocalOrder] Remove check of discontinuous map task ids (#354)
### What changes were proposed in this pull request?
Remove check of discontinuous map task ids

### Why are the changes needed?
In the [startMapIndex, endMapIndex) sequence, the index is not the map task id, which only indicate the range of the succeed map task ids set. 

So that means. if the range is [0, 5), it only indicate the number of tasks is 5. And the task id list may be as follows:

`[task-1, task-3, task-9, task-15, task-16]`


### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
1. UTs