[ISSUE-228][FEATURE] Add a period local storage cleaner thread (#357)
### What changes were proposed in this pull request?
Add a period local storage cleaner thread to check the leak shuffle data on disks

### Why are the changes needed?
Because we don't use strict lock strategy, we will delete some shuffle data normally. If we use strict lock strategy, the performance will be bad. So we want to add a period local storage cleaner thread. The thread examine which application is in the disk and judge whether the app exists the memory data structure ShuffleTaskManager#shuffleTaskInfos. If memory data structure don't exist, we will delete the data in the disk, with issue #282 

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
tests were added

Co-authored-by: wangshifa <wangshifa@58.com>