[ISSUE-468] Put unavailable servers to the end of the list when sending shuffle data (#470)
### What changes were proposed in this pull request?
Put unavailable shuffle servers to the end of the server list when sending shuffle data if replica=1.

### Why are the changes needed?

If we use multiple replicas and the first shuffle server becomes unavailable, sending data will take a lot of time. Because the client will always send to the first shuffle server firstly. #468 

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT