[ISSUE-529][operator] fix error kind of ownerreference (#540)
### What changes were proposed in this pull request?
Fix #529 

### Why are the changes needed?
This fix #529. 
After this fix, garbage collection in k8s cluster can be completed normally.

### Does this PR introduce _any_ user-facing change?
This is a bug fix, after this pr.  Child resource created by rss could be gced automatically.

### How was this patch tested?
After the rss object is deleted, see if the rss-dependent objects (such as services and statefulsets) are cleaned up with it to verify that the function is normal.