[#665] feat(client): keep consistent with vanilla spark when key or value is null (#666)
### What changes were proposed in this pull request?

keep consistent with vanilla spark when key or value is null

### Why are the changes needed?

Fix: #665 

The PR of https://github.com/apache/incubator-uniffle/pull/296 has fixed the bug of NPE of value. But in some cases of production, the key also could be null. So this corner case also should keep consistent with vanilla spark.

And to prevent this logic from reverting, the corresponding tests have been attached.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

1. Unit tests