[#674] feat(docker): use JDK11 as the default java version in Dockerfile (#683)
### What changes were proposed in this pull request?

Use JDK11 as the default java version in Dockerfile when deploying shuffle servers on K8s.

### Why are the changes needed?

As the GC problems mentioned in #674, I upgraded the JVM version from JDK8 to JDK11 for shuffle server. And after benchmark, I found the effect of the upgrade is remarkable. And it also works well with Hadoop3.1.1

Based on above practice, it's better to use JDK11 as the default java version to improve stability

### Does this PR introduce _any_ user-facing change?

Yes

### How was this patch tested?

Have tested in production env.

Co-authored-by: zhangjunfan <zhangjunfan@qiyi.com>