fix(client): disable spark memory spill (#844)
### What changes were proposed in this pull request?

Disable the memory spill operation.

### Why are the changes needed?

In #714 , the memory spill is introduced to solve the dead lock. For a pity,
these part code should be handled carefully, including concurrency and data
consistency, like the fix PR #811 . And this part has bugs and I will fix these in
the next days. 

Currently, I want to revert the PR #714. But the partial refactor of #714 is still meaningful. So
I submit this PR to disable the memory spill.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Don't need