[MINOR] refactor: Reduce the usage of memory in the `ShuffleWriter` (#877)
### What changes were proposed in this pull request?
We should immediately remove the `WriteBuffer` after  we use the method `createShuffleBlock` to create `ShuffeBlock`. Otherwise we will have the compressed data and uncompressed data at the same time in the memory. JVM's GC can't collect the memory of uncompressed data, because the data have the reference in the map `buffers`.

### Why are the changes needed?
Reduce the usage of memory.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
CI passed.