[MINOR] improvement: use mvn wrapper for builds (#1345)
### What changes were proposed in this pull request?

Introducing maven wrapper (version: `3.9.6`). While working on building and deploying `uniffle` it took me considerably long to pin the right version of maven. Looking at: #685 it seems that I wasn't the only person who had this problem (the service docs aren't clear about which maven version works). I think that introducing maven wrapper and pinning its version to the working configuration will reduce the friction for new contributors and users of the shuffle service.

### Why are the changes needed?

Pinning down the maven version reduces friction for the project setup (one doesn't need to ensure a specific version of maven is installed) and builds stability (the build is more resilient to changes in the build environments).

Fix: #685

### Does this PR introduce _any_ user-facing change?

We're moving from `mvn` to `./mvw` command to execute build-related tasks.

### How was this patch tested?

I've ensured that:
* `build.sh` and `build_distribution.sh` build valid `uniffle` distribution
*  all `./mvnw*` commands listed in the README files are executed without errors.