[MINOR] improvement: use mvn wrapper in CI builds. (#1351)
### What changes were proposed in this pull request?

This is a follow-up to #1345. The change will (hopefully) invoke `./mvnw` (wrapper) instead of the `mvn` command in CI builds, making the local development builds and CI builds equivalent to the applied `maven` version. 

### Why are the changes needed?

It's prudent to have the development builds as close to the ones executed by CI/CD system as possible.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

I believe having a PR builder pass should suffice.