[MINOR] refactor: Avoid unnecessary recursion (#1441)
### What changes were proposed in this pull request?
Change the recursion of ShuffleReadClientImpl.readShuffleBlockData into an outer while(true) loop. Diff is best compared with whitespaces ignored.

### Why are the changes needed?
The recursion in ShuffleReadClientImpl.readShuffleBlockData is not needed because the existing return statements also terminate the introduced outer loop. The recursion depth is limited by the number of consecutive empty segments.

### Does this PR introduce any user-facing change?
No.

### How was this patch tested?
Not tested.