[MINOR] refactor: avoid unnecessary bitmap clone and AND (#1442)
### What changes were proposed in this pull request?
Add note that processedBlockIds can be a superset of blockIdBitmap. Add quick check if processedBlockIds is equal to blockIdBitmap. In that case, there is no need to clone and AND the bitmaps.

### Why are the changes needed?
The logic is equivalent but more performant when processedBlockIds is not a superset of blockIdBitmap (when no other blocks than the blockIdBitmap are retrieved).

### Does this PR introduce any user-facing change?
No.

### How was this patch tested?
Existing tests in ShuffleReadClientImplTest cover this.