[#1355] fix(client): Netty client will leak when decoding responses (#1455)
### What changes were proposed in this pull request?

The current code logic is that the `ByteBuf` is only released when `msg.body() == null`. However, when `msg.body != null`, `msg.body().byteBuf()` returns a `NettyManagedBuffer.EMPTY_BUFFER`, and the `ByteBuf` is not released in this case, resulting in a memory leak issue every time decoding an RPC response from ShuffleServer.
Over time, if the Spark Job runs long enough and there are enough requests, it will eventually cause a significant memory leak on the client side (Spark Executor).
The modifications to the other code are mainly for readability and enhanced protection, and will not cause any side effects.

### Why are the changes needed?

To fix the memory leak issue in the Netty client when decoding RPC responses.
For [#1359](https://github.com/apache/incubator-uniffle/issues/1359)

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Rerun successfully and no more Netty leak logs.