feat(docker): Add example docker compose Uniffle/Spark cluster (#1532)
### What changes were proposed in this pull request?
Adds code to spin up an example Uniffle/Spark docker cluster using docker compose. This is used by the CI to test the example cluster setup.

### Why are the changes needed?
This setup has a smaller footprint than the existing kubernetes example in `deploy/kubernetes/integration-test/e2e/README.md`, which is not trivial to setup. The new example only requires Docker to be installed, and can be spun up via

    ./deploy/docker-compose/build.sh
    docker compose -f deploy/docker-compose/docker-compose.yml up

The Uniffle and Spark cluster can be used to interactively test Spark with Uniffle.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Manually and CI tested.