[#1591] feat(spark): Support Spark 3.5.1 (#1592)
### What changes were proposed in this pull request?

Support Spark 3.5.1

The abstract class `TaskContext` introduced a new method `isFailed()` after Spark version 3.5.1. Due to this addition, it is not possible to compile the `MockTaskContext` class using Spark 3.5.1 without modifying the code. Consequently, we also added the `isFailed()` method to `MockTaskContext`. However, to maintain compatibility with versions of Spark prior to 3.5.1, we did not annotate this method with `@Override`.

Compiling old code with Spark 3.5.1 results in the following error: 
```error: MockTaskContext is not abstract and does not override abstract method isFailed() in TaskContext.```
Additionally, adding the `@Override` annotation leads to another error: 
```error: method does not override or implement a method from a supertype.```

### Why are the changes needed?

Fix: #1591 

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests.