[MINOR] refactor: Calling lock() method outside try block to avoid unnecessary errors (#1590)
### What changes were proposed in this pull request?

Calling `lock()` method outside try block to avoid unnecessary errors

### Why are the changes needed?

In general, the `lock()` method should be placed outside the try block. The reason is that if the `lock()` method throws an exception, it indicates that the lock was not acquired, so there is no need to attempt to release it in the finally block. If the `lock()` method is placed within the try block and it throws an exception, the finally block will still be executed and attempt to release a lock that was never acquired, leading to errors.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing UTs.