[#1626] fix(server): Remove the meaningless eventOfUnderStorageManagers cache (#1627)
### What changes were proposed in this pull request?

Remove the meaningless eventOfUnderStorageManagers.

### Why are the changes needed?

Fix https://github.com/apache/incubator-uniffle/issues/1626 & https://github.com/apache/incubator-uniffle/issues/1620.
It's also a follow-up PR for https://github.com/apache/incubator-uniffle/pull/383.

This cache only makes sense when the event retries after a failure. However, after the event fails, it is not appropriate to continue taking the original storageManager from the cache(because events usually fail due to high IO pressure or disk damage or disk full). In this case, the cache seems to be meaningless, so there is a contradiction here, we should remove it.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Tested in our env