[#1745] feat(remote merge): Introduce a common serializer. (#1916)
### What changes were proposed in this pull request?


### Why are the changes needed?

Since merge needs to be performed on the ShuffleServer side, and serialization/deserialization may differ between versions of the compute framework. To avoid server-side processing of different versions of serialization/deserialization, unified serializer that is independent of the computing framework needs to be extracted. 
For now, only one serializers are supported, just WritableSerializer. WritableSerializer is used to handle the org.apache.hadoop.io.Writable interface, used in the MR and TEZ frameworks. 

> Note: KryoSerializer will be introduced when spark is supported.

Fix: #1745

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

unit test and test in cluster.