[#1401] feat(dashboard): Add mock data to the Dashboard front end. (#1989)
### What changes were proposed in this pull request?

When the Dashboard front-end development needs to show some effects, it is necessary to start the backend to provide the corresponding interface to see the effect. After the mock data is added, you can see the page effect by intercepting the network request and directly using the mock data of the front-end.

### Why are the changes needed?

Fix: #1401 

### Does this PR introduce _any_ user-facing change?

Yes.

### How was this patch tested?

Existing UTs.