[MINOR] improve(server): Add return value to server rpc audit log (#2007)
### What changes were proposed in this pull request?

Add return value for server rpc audit log

### Why are the changes needed?

With the return value, we can know clear about the effect of the rpc call.

E.g. A bundle of `requireBuffer` calls appear in the rpc audit log, then followed some `sendShuffleData` rpc call with `requireBufferId`, but we don't know which `requireBuffer` call match the sendShuffleData call.

With this PR, we record the requireBufferId also which generated by `requireBuffer` call, and search the sendShuffleData call with the `requireBufferId`, this can be possible.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Test Locally.