[#2087] feat(operator): support closing leader election of webhook of rss-operator (#2089)
### What changes were proposed in this pull request?

In this PR, we set a flag to open or close leader election of webhook of rss-operator.

We set the default value to false because leader election is only involved when synchronizing configuration objects used by webhook. Even if leader election is disabled, there will be no conflicts because the configuration objects rarely change. We plan to consider removing leader election completely in the future.

### Why are the changes needed?

Fix: #2087 

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Added ut or Manually verified.