[#2060] fix(server): Fix memory leaks when reaching memory limit (#2058)
### What changes were proposed in this pull request?

Fix shuffle server memory leaks when reaching memory limit.

### Why are the changes needed?
When enabling Netty, on the shuffle server side, Netty will allocate memory for SEND_SHUFFLE_DATA_REQUEST request. However, when the memory limit is reached, an `OutOfDirectMemoryError` will be thrown and the decoding process for this message will fail. This will cause the bytebuf allocated successfully in the previous batch in this message to not be released, resulting in memory leaks.

Fix: https://github.com/apache/incubator-uniffle/issues/2060

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?
Existing UTs